refactored project selection to it's own file and conponent from project.tsx
